requirements.txt
streamlit==1.32.0
pandas==2.2.1
matplotlib==3.8.2
plotly==5.19.0
sqlalchemy==2.0.27
python-dotenv==1.0.1
pytest==8.0.2
black==24.2.0
flake8==7.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
scikit-learn==1.3.2
bcrypt==4.0.1
python-multipart==0.0.9
openpyxl==3.1.2
python-dateutil==2.8.2
pytz==2024.1